requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-filer==2.0.2
django-grappelli==2.15.1
django-js-asset==1.2.2
django-material-admin==1.8.5
django-mptt==0.12.0
django-polymorphic==3.0.0
easy-thumbnails==2.7.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.1.2
